% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AtCatalog}
\alias{AtCatalog}
\title{A list of TF-target genes (TG) in Arabidopsis thaliana.}
\value{
None
}
\description{
A catalogue of annotated target genes in Arabidopsis thaliana
}
\details{
The original regulatory network datasets were downloaded from AtRegNet
- AGRIS Arabidopsis
(http://arabidopsis.med.ohio-state.edu/moreNetwork.html)
and ATRM (http://atrm.cbi.pku.edu.cn/), respectively. Then we integrated
them based on AGI code.

The catalogues version is 0.0.1.
}
\references{
Yilmaz A et al. NAR 39:D1118-1122 (2011)
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
