% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATRM}
\alias{ATRM}
\title{A list of ATRM TF-target genes (TG) in Arabidopsis thaliana.}
\value{
None
}
\description{
A catalogue of annotated target genes in Arabidopsis thaliana based on ATRM
}
\details{
The original regulatory network datasets were downloaded from
ATRM (http://atrm.cbi.pku.edu.cn/).
}
\references{
Jin JP et al. Mol Bio Evol 32:1767-1773 (2015)
}
\keyword{datasets}
