% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEQvalue.R
\name{calculateEvalue}
\alias{calculateEvalue}
\title{calculates E-value}
\usage{
calculateEvalue(df, TFs)
}
\arguments{
\item{df}{a data frame containng p-values}

\item{TFs}{a character vector of transcription factor}
}
\value{
data.frame
}
\description{
This function calculates E-value based on .
}
\examples{
data(DEGs)
data(catalog)

tftg <- extractTFTG(DEGs, catalog)
TFs <- tftg$TFs

p.value <- runif(10)/(1:10)
df <- data.frame(p.value = p.value)
res <- calculateEvalue(df, TFs)

}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
