% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEQvalue.R
\name{calculateFDRBH}
\alias{calculateFDRBH}
\title{calculates FDR by Benjamini and Hochberg method}
\usage{
calculateFDRBH(df)
}
\arguments{
\item{df}{a data frame containng p-values}
}
\value{
data.frame
}
\description{
This function calculates FDR based on BH.
}
\examples{
p.value <- runif(10)/(1:10)
df <- data.frame(p.value = p.value)
res <- calculateFDRBH(df)

}
\references{
Benjamini Y and Hochberg Y, J Roy Stat Soc B 57: 289?300 (1995)
}
\author{
Atsushi Fukushima
}
