% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalog}
\alias{catalog}
\title{A list of TF-target genes (TG) in TFactS.}
\value{
None
}
\description{
A catalogue of annotated target genes based on TFactS.
}
\details{
This was downloaded from TFactS (http://www.tfacts.org/).

The catalogues version is 2
(http://www.tfacts.org/TFactS-new/TFactS-v2/tfacts/data/Catalogues.xls).
}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
