% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGs}
\alias{DEGs}
\title{An example list of DEGs from DEMO data in TFactS.}
\value{
None
}
\description{
An example list of differentially expressed genes (DEGs).
}
\details{
This was from DEMO data in TFactS (http://www.tfacts.org/).
The list corresponds to 18 up-regulated genes.
}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
