% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGs39}
\alias{DEGs39}
\title{An example list of 39 DEGs from DEMO data in TFactS.}
\value{
None
}
\description{
A list of 39 differentially expressed genes (DEGs).
}
\details{
This was from DEMO data in TFactS (http://www.tfacts.org/). The list
corresponds to 18 up-regulated genes and 21 down-regulated genes.
}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
