% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.df}
\alias{example.df}
\title{An example result of enrichment analysis}
\value{
None
}
\description{
An example result of enrichment analysis
}
\details{
The size of data frame is 345 TFs and 6 columns.
}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
