% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.list}
\alias{example.list}
\title{An example list of result of enrichment analysis}
\value{
None
}
\description{
An example list of result of enrichment analysis
}
\details{
The length of the list is 345 (TFs).
}
\references{
Essaghir A et al. Nucleic Acids Res. 2010 Jun;38(11):e120.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
