% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatET}
\alias{formatET}
\title{formats the result of enrichment test (ET)}
\usage{
formatET(list)
}
\arguments{
\item{list}{a list of the result of enrichment test (ncol = 6)}
}
\value{
data.frame
}
\description{
This function formats the result of enrichment test (ET)
based on Fisher's exact test
}
\examples{
data(example.list)
res <- formatET(example.list)

}
\author{
Atsushi Fukushima
}
