% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatRC}
\alias{formatRC}
\title{formats the result of Random Control (RC)}
\usage{
formatRC(df, list, nRep)
}
\arguments{
\item{df}{a data frame of ET including E-values, FDR-BH, and Q-values}

\item{list}{a list of the result of RC (ncol = 2)}

\item{nRep}{the number of random selections (negative control)}
}
\value{
data.frame
}
\description{
This function formats the result of Random Control (RC)
with random simulation based on Fisher's exact test
}
\examples{
data(example.df)
data(example.list)
nRep <- 100
res <- formatRC(example.df, example.list, nRep)

}
\author{
Atsushi Fukushima
}
