% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenesUp_SH1H}
\alias{GenesUp_SH1H}
\title{An example list of 74 DEGs from AtGenExpress stress dataset.}
\value{
None
}
\description{
A list of 74 differentially expressed genes (DEGs) from AtGenExpress
}
\details{
This was from DEMO data in the Supplemental data (Table S4. Genes
regulated by UV-B light stress in roots and shoots; filename:
TPJ_3052_sm_TabS4.xls) of the paper by Killian et al. TPJ (2008).
The list corresponds to 74 genes up-regulated in roots and shoots 1 hour
after onset of UV-B stress treatment.
}
\references{
Killian J et al. Plant J. 2007 Apr;50(2):347-363.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
